fix: extract ToastIcon to new file
